Finish gpu_utils_polars consolidation and rename core pipeline modules (TASK-51/TASK-52)
TASK-51: move gpu_utils_polars.py's Parquet/SQLite file-discovery functions
into stats_data.py and delete the now-empty module; fix scripts/plot_gpu_availability.py,
which was silently broken against current data since it only knew the pre-migration
SQLite-only discovery path.

TASK-52 (naming pass only): rename stats_data.py -> read_data.py,
stats_calculations.py -> classify_slots.py, stats_reporting.py -> reporting.py,
device_name_mappings.py -> devices.py. Delete gpu_utils.py, splitting its content
by responsibility: host-config loading into read_data.py, filter_df/filter_df_enhanced
and related classification state into classify_slots.py, and the two display-only
helpers into reporting.py. Update every caller, both Dockerfiles, and the test suite
accordingly.
