+++ analysis/analyze_task7_troubleshoot.py
@@ -14,7 +14,8 @@ from collections import defaultdict
 import pandas as pd
 
 # Import the exact filter function and host exclusion logic from usage_stats.py
-from gpu_utils import filter_df, load_host_exclusions
+from classify_slots import filter_df
+from read_data import load_host_exclusions
 
 
 def should_include_device(device_name: str, include_all_devices: bool = False) -> bool:
@@ -32,10 +33,10 @@ def analyze_gpu_categories(db_path: str, include_all_devices: bool = False) -> d
     """Analyze GPU categories and return detailed breakdown."""
     # Set up host exclusions like usage_stats.py does
     global HOST_EXCLUSIONS, FILTERED_HOSTS_INFO
-    import gpu_utils
+    import classify_slots
 
-    gpu_utils.HOST_EXCLUSIONS = load_host_exclusions(None, "masked_hosts.yaml")
-    gpu_utils.FILTERED_HOSTS_INFO = []
+    classify_slots.HOST_EXCLUSIONS = load_host_exclusions(None, "masked_hosts.yaml")
+    classify_slots.FILTERED_HOSTS_INFO = []
 
     # Connect to database and load data
     conn = sqlite3.connect(db_path)
