+++ scripts/plot_usage_stats.py
@@ -16,8 +16,8 @@ import pandas as pd
 import typer
 
 # Import functions from usage_stats
-from stats_calculations import calculate_allocation_usage_by_device, calculate_time_series_usage
-from stats_data import get_time_filtered_data
+from classify_slots import calculate_allocation_usage_by_device, calculate_time_series_usage
+from read_data import get_time_filtered_data
 
 
 def filter_by_gpu_model(df: pd.DataFrame, gpu_model: str) -> pd.DataFrame:
