+++ tests/test_usage_stats.py
@@ -18,13 +18,13 @@ import pytest
 
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
-from gpu_utils import filter_df
-from stats_calculations import (
+from classify_slots import (
     calculate_allocation_usage,
     calculate_allocation_usage_by_device,
     calculate_time_series_usage,
+    filter_df,
 )
-from stats_data import get_time_filtered_data
+from read_data import get_time_filtered_data
 from usage_stats import run_analysis
